Fix cannot assign null to property Interceptor::$client
We have recently added a type hint to the `Interceptor::$client` property. This will result in `TypeError(code: 0): Cannot assign null to property Grphp\StatsD\Interceptor::$client of type Domnikl\Statsd\Client` error if the client option isn't correctly set, when it actually should fallback onto instantiating one. Fix the problem by not assigning the `null` value to the `Interceptor::$client` propery if the `client` option is not set.